fix(review): address F01 - extract duplicated chain-walking helpers shared with S8780 into helpers/expect-chain.ts
